requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
anything==0.2.1
asgiref==3.7.2
asttokens==2.4.1
attrs==24.3.0
autobahn==20.12.3
Automat==22.10.0
beautifulsoup4==4.12.3
beautifultable==1.1.0
bidict==0.23.1
black==23.3.0
bleak==0.22.3
blinker==1.5
bottle==0.13.2
bottle-websocket==0.2.9
bs4==0.0.1
build==1.2.1
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
constantly==15.1.0
contourpy==1.3.1
cryptography==42.0.7
cycler==0.12.1
debugpy==1.8.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.6
distro==1.9.0
Django==5.0.7
django-classy-tags==4.1.0
django-filter==2.4.0
django-sekizai==2.0.0
djangorestframework==3.14.0
docutils==0.21.2
Eel==0.18.1
et-xmlfile==1.1.0
executing==2.0.1
ez_setup==0.9
fastapi==0.110.3
filelock==3.15.4
Flask==2.1.1
Flask-Mail==0.9.1
flatbuffers==24.3.25
fonttools==4.55.3
fpdf2==2.7.4
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
gunicorn==20.1.0
h11==0.14.0
h2==2.6.2
hpack==3.0.0
httpcore==1.0.5
httpx==0.27.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==21.0.0
idna==3.10
imageio==2.34.2
importlib_metadata==8.2.0
incremental==22.10.0
inflect==6.0.4
ipykernel==6.29.5
ipython==8.26.0
iso8601==2.1.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.2
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==25.3.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lml==0.1.0
lunr==0.6.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mock==5.0.2
model-mommy==2.0.0
more-itertools==10.3.0
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.18
numba==0.60.0
numpy
odfpy==1.4.1
opencv-python==4.6.0.66
openmeteo_requests==1.3.0
openmeteo_sdk==1.18.6
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parameterized==0.8.1
parso==0.8.4
pathspec==0.11.1
patsy==0.5.6
pdfkit==1.0.0
pillow==11.1.0
pipdeptree==2.16.2
pixelate==1.0.0
pkginfo==1.10.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
protobuf==5.29.3
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-ods==0.6.0
Pygments==2.18.0
pymud==0.20.0
pyOpenSSL==24.1.0
pyparsing==3.2.1
pyperclip==1.9.0
pyproj==3.7.1
pyproject_hooks==1.1.0
pyserial==3.5
pysondb==1.6.7
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pythondb==2.2
pytz==2024.2
pywin32==306
pywin32-ctypes==0.2.2
pyxelate @ git+https://github.com/sedthh/pyxelate.git@fed220be1f97014d9d8607547cc1d4340044f7d9
PyYAML==6.0.1
pyzmq==26.0.3
ratelim==0.1.6
readme_renderer==44.0
redis==5.0.8
reflex==0.5.8
reflex-hosting-cli==0.1.13
requests==2.32.3
requests-cache==1.2.1
requests-toolbelt==1.0.0
retry-requests==2.0.0
rfc3986==2.0.0
rich==13.7.1
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
SecretStorage==3.3.3
setuptools==75.6.0
setuptools-scm==7.0.5
shapely==2.0.7
shellingham==1.5.4
simple-file-checksum==1.2.2
simple-websocket==1.0.0
simpleeval==0.9.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlmodel==0.0.21
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
starlette-admin==0.14.1
statsmodels==0.14.2
tabulate==0.9.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.7.24
tomli==2.0.1
tomlkit==0.13.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
twine==5.1.1
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.3.0
uvicorn==0.30.5
virtualenv==20.16.6
waitress==3.0.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.5
websockets==12.0
Werkzeug==2.1.1
wheel==0.43.0
winrt-runtime==2.3.0
winrt-Windows.Devices.Bluetooth==2.3.0
winrt-Windows.Devices.Bluetooth.Advertisement==2.3.0
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==2.3.0
winrt-Windows.Devices.Enumeration==2.3.0
winrt-Windows.Foundation==2.3.0
winrt-Windows.Foundation.Collections==2.3.0
winrt-Windows.Storage.Streams==2.3.0
wkhtmltopdf==0.2
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.2
zipp==3.19.2
zope.event==5.0
zope.interface==7.2
